[refactor] LottoNumberCache를 제거하고 LottoNumber에 정적 팩토리 메소드를 추가
