+++ src/main/java/lotto/LottoApplication.java
@@ -3,11 +3,9 @@ package lotto;
 import lotto.controller.LottoMachine;
 import lotto.domain.WinningLotto;
 import lotto.domain.Lotto;
-import lotto.domain.LottoNumber;
 import lotto.domain.LottoResult;
 import lotto.ui.InputView;
 import lotto.ui.ResultView;
-import lotto.util.LottoNumberParser;
 
 public class LottoApplication {
     public static void main(String[] args) {
