+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -7,6 +7,16 @@ import static org.assertj.core.api.Assertions.*;
 public class LottoNumberTest {
     @Test
     public void 생성() {
-        assertThat(new LottoNumber(1)).isEqualTo(new LottoNumber(1));
+        assertThat(LottoNumber.get(1)).isEqualTo(LottoNumber.get("1"));
+    }
+
+    @Test
+    public void 캐싱() {
+        assertThat(LottoNumber.get(1)).isEqualTo(LottoNumber.get(1));
+    }
+
+    @Test
+    public void 유효하지_않은_숫자() {
+        assertThatThrownBy(() -> LottoNumber.get(46)).isInstanceOf(IllegalArgumentException.class);
     }
 }
