+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -27,7 +27,7 @@ public class WinningLottoTest {
     public void 일치하는_숫자의_개수(String numbers, LottoRank expected) {
         LottoNumbers lottoNumbers = new LottoNumbers(numbers);
         LottoNumbers winningLottoNumbers = new LottoNumbers(1, 2, 3, 4, 5, 6);
-        LottoNumber bonusNumber = new LottoNumber(13);
+        LottoNumber bonusNumber = LottoNumber.get(13);
         WinningLotto winningLotto = new WinningLotto(winningLottoNumbers, bonusNumber);
 
         assertThat(winningLotto.getMatchedRank(lottoNumbers)).isEqualTo(expected);
